Testing something

+++ Jenkinsfile
@@ -13,10 +13,8 @@ pipeline {
             steps {
                 sh """
                 pip install --user setuptools-scm more-itertools==5.0.0 tox twine==1.15.0
-                export ftp_proxy="$http_proxy"
                 export http_proxy="http://cloud-proxy:3128"
                 export https_proxy="http://cloud-proxy:3128"
-                echo $ftp_proxy
                 tox
                 """
             }
@@ -28,14 +26,19 @@ pipeline {
                 TWINE_PASSWORD = credentials('twine_password')
             }
             steps {
-                sh """
-                python setup.py --version
-                rm -rf dist/*
-                python setup.py sdist bdist_wheel
-                twine upload -r gdcsnapshots dist/*
-                """
+                script {
+                    if (env.BRANCH_NAME == 'feat/setuptools_scp') {
+                        sh """
+                        python setup.py --version
+                        rm -rf dist/*
+                        python setup.py sdist bdist_wheel
+                        twine upload -r gdcsnapshots dist/*
+                        """
+                    } else {
+                        echo 'I execute elsewhere'
+                    }
+                }
             }
         }
     }
 }
-
